@@ -16,27 +16,27 @@
         "test": "ava"
     },
     "dependencies": {
-        "autoprefixer": "^9.0",
-        "chokidar": "^3.0.0",
-        "csso": "^3.5.1",
-        "fs-extra": "^8.0.1",
+        "autoprefixer": "^9.7.3",
+        "chokidar": "^3.3.1",
+        "csso": "^4.0.2",
+        "fs-extra": "^8.1.0",
         "kleur": "^3.0.1",
-        "node-sass": "^4.11.0",
-        "postcss": "^7.0",
+        "node-sass": "^4.13.0",
+        "postcss": "^7.0.25",
         "postcss-reporter": "^6.0.1",
         "pretty-hrtime": "^1.0.3",
-        "stylelint": "^10.0.1"
+        "stylelint": "^12.0.0"
     },
     "devDependencies": {
         "@types/chokidar": "^2.1.3",
-        "@types/fs-extra": "^7.0.0",
-        "@types/node": "^12.0.2",
+        "@types/fs-extra": "^8.0.1",
+        "@types/node": "^12.12.18",
+        "@types/node-sass": "^4.11.0",
         "@types/sass": "^1.16.0",
-        "@types/stylelint": "^9.4.1",
-        "ava": "^1.4.1",
-        "glob": "^7.1.4",
-        "source-map": "^0.7.3",
-        "typescript": "^3.4.5"
+        "@types/stylelint": "^9.10.1",
+        "ava": "^2.4.0",
+        "glob": "^7.1.6",
+        "typescript": "^3.7.3"
     },
     "engines": {
         "node": ">=8.0"
